move swervedrive command to drivetrain folder
