+++ src/main/java/frc/robot/RobotContainer.java
@@ -17,10 +17,10 @@ import edu.wpi.first.wpilibj2.command.ScheduleCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.commands.routines.IntakeNote;
-import frc.robot.commands.SwerveDrive;
 import frc.robot.commands.climber.ClimbDown;
 import frc.robot.commands.climber.ClimbUp;
 import frc.robot.commands.drivetrain.AutoAlignToSpeaker;
+import frc.robot.commands.drivetrain.SwerveDrive;
 import frc.robot.commands.intake.FeedNote;
 import frc.robot.commands.intake.ForceSlowGrabNote;
 import frc.robot.commands.intake.GrabNote;
